Create seperate sum test file
